Use macro.
