+++ crates/core/src/tool.rs
@@ -30,6 +30,8 @@ pub trait Tool<'tool>:
 
     fn get_manifest(&self) -> Result<&Manifest, ProtoError>;
 
+    fn get_manifest_mut(&mut self) -> Result<&mut Manifest, ProtoError>;
+
     fn get_manifest_path(&self) -> PathBuf {
         self.get_tool_dir().join(MANIFEST_NAME)
     }
@@ -164,4 +166,33 @@ pub trait Tool<'tool>:
     async fn after_teardown(&mut self) -> Result<(), ProtoError> {
         Ok(())
     }
+}
+
+#[macro_export]
+macro_rules! impl_tool {
+    ($tool:ident) => {
+        impl Tool<'_> for $tool {
+            fn as_any(&self) -> &dyn Any {
+                self
+            }
+
+            fn get_manifest(&self) -> Result<&Manifest, ProtoError> {
+                self.manifest
+                    .get_or_try_init(|| Manifest::load(self.get_manifest_path()))
+            }
+
+            fn get_manifest_mut(&mut self) -> Result<&mut Manifest, ProtoError> {
+                {
+                    // Ensure that the manifest has been initialized
+                    self.get_manifest()?;
+                }
+
+                Ok(self.manifest.get_mut().unwrap())
+            }
+
+            fn get_tool_dir(&self) -> &Path {
+                &self.base_dir
+            }
+        }
+    };
 }
