+++ crates/node/src/depman/mod.rs
@@ -7,7 +7,7 @@ mod shim;
 mod verify;
 
 use once_cell::sync::OnceCell;
-use proto_core::{Describable, Manifest, Proto, ProtoError, Tool};
+use proto_core::{impl_tool, Describable, Manifest, Proto, ProtoError, Tool};
 // use resolve::NDMVersionDist;
 use std::{
     any::Any,
@@ -80,17 +80,4 @@ impl Describable<'_> for NodeDependencyManager {
     }
 }
 
-impl Tool<'_> for NodeDependencyManager {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn get_manifest(&self) -> Result<&Manifest, ProtoError> {
-        self.manifest
-            .get_or_try_init(|| Manifest::load(self.get_manifest_path()))
-    }
-
-    fn get_tool_dir(&self) -> &Path {
-        &self.base_dir
-    }
-}
+impl_tool!(NodeDependencyManager);
