+++ crates/rust/src/lib.rs
@@ -7,7 +7,7 @@ mod shim;
 mod verify;
 
 use once_cell::sync::OnceCell;
-use proto_core::{Describable, Manifest, Proto, ProtoError, Tool};
+use proto_core::{impl_tool, Describable, Manifest, Proto, ProtoError, Tool};
 use std::{
     any::Any,
     path::{Path, PathBuf},
@@ -47,17 +47,4 @@ impl Describable<'_> for RustLanguage {
     }
 }
 
-impl Tool<'_> for RustLanguage {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn get_manifest(&self) -> Result<&Manifest, ProtoError> {
-        self.manifest
-            .get_or_try_init(|| Manifest::load(self.get_manifest_path()))
-    }
-
-    fn get_tool_dir(&self) -> &Path {
-        &self.base_dir
-    }
-}
+impl_tool!(RustLanguage);
